requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1

sqlalchemy==2.0.35

pydantic-settings==2.5.2
python-dotenv==1.0.1

pytesseract==0.3.13

openai==1.51.0

pymupdf==1.24.10

Pillow==10.4.0
scikit-image==0.24.0
numpy==1.26.4

httpx==0.27.2

watchdog==5.0.3
